pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "User Management + API Gateway Backend"
authors = ["Mikhail Liamets <m.liamets@jacobs-university.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-ninja = "^0.21.0"
psycopg2 = "^2.9.6"
mypy = "^1.2.0"
django-stubs = "^1.16.0"
pre-commit = "^3.2.2"
pyupgrade = "^3.3.1"
autoflake = "^2.0.2"
django-filter = "^23.1"
django-extensions = "^3.2.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.2"
django-filter-stubs = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
mypy_path = "core"
python_version = 3.9
pretty = true
show_traceback = true
color_output = true
plugins = [
"pydantic.mypy",
"mypy_django_plugin.main"
]

[tool.django-stubs]
django_settings_module = "core.settings"
